Generate first alembic revisision
