+++ alembic/versions/a1e3760a814b_first_revision.py
@@ -1,0 +1,37 @@
+"""First revision
+
+Revision ID: a1e3760a814b
+Revises:
+Create Date: 2022-02-03 19:05:48.223249
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'a1e3760a814b'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('country',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('iso_code', sa.String(), nullable=True),
+    sa.Column('name', sa.String(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_country_id'), 'country', ['id'], unique=False)
+    op.create_index(op.f('ix_country_iso_code'), 'country', ['iso_code'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_country_iso_code'), table_name='country')
+    op.drop_index(op.f('ix_country_id'), table_name='country')
+    op.drop_table('country')
+    # ### end Alembic commands ###
