--- alembic/versions/d4867f3a4c0a_first_revision.py
@@ -1,59 +1,0 @@
-"""First revision
-
-Revision ID: d4867f3a4c0a
-Revises:
-Create Date: 2019-04-17 13:53:32.978401
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "d4867f3a4c0a"
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "user",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("full_name", sa.String(), nullable=True),
-        sa.Column("email", sa.String(), nullable=True),
-        sa.Column("hashed_password", sa.String(), nullable=True),
-        sa.Column("is_active", sa.Boolean(), nullable=True),
-        sa.Column("is_superuser", sa.Boolean(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_user_email"), "user", ["email"], unique=True)
-    op.create_index(op.f("ix_user_full_name"), "user", ["full_name"], unique=False)
-    op.create_index(op.f("ix_user_id"), "user", ["id"], unique=False)
-    op.create_table(
-        "item",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("title", sa.String(), nullable=True),
-        sa.Column("description", sa.String(), nullable=True),
-        sa.Column("owner_id", sa.Integer(), nullable=True),
-        sa.ForeignKeyConstraint(["owner_id"], ["user.id"],),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_item_description"), "item", ["description"], unique=False)
-    op.create_index(op.f("ix_item_id"), "item", ["id"], unique=False)
-    op.create_index(op.f("ix_item_title"), "item", ["title"], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_item_title"), table_name="item")
-    op.drop_index(op.f("ix_item_id"), table_name="item")
-    op.drop_index(op.f("ix_item_description"), table_name="item")
-    op.drop_table("item")
-    op.drop_index(op.f("ix_user_id"), table_name="user")
-    op.drop_index(op.f("ix_user_full_name"), table_name="user")
-    op.drop_index(op.f("ix_user_email"), table_name="user")
-    op.drop_table("user")
-    # ### end Alembic commands ###
